Print a Fitted gpAim Object

Prints a concise summary of a gpAim fit to the console, reporting the number of genotyped lines and markers, the computational path used ("vm" or "mbf"), estimated genetic and residual variance components, narrow-sense heritability, and the range of GEBVs.

Usage

# S3 method for class 'gpAim'
print(x, ...)

Arguments

x

A fitted object of class "gpAim".

...

Currently unused.

Value

x invisibly.

See also

Examples

if (FALSE) { # \dontrun{
print(gp.fit)
} # }